Add Message handling, own sessions handling, move api key & secret to appsettings, view restiction and more
### BackOffice:
Add: Own sessions handling (Model/Sessions and Model/Session)
Add: Messages handling through sessions (Model/MessagesHandler and Helper/MessagesHelper)
Add: View restictions for only login or logout visitors
Update: Actions in the Controllers
Update: Views
Update: _Layout, add gobal support for messages
Move: Api key and secret to appsettings
Move: View Users/Users to Users/Index
Remove: Model/ApiInformation

### RestApiClient:
Change: TicketApi add handling of InternalServerError
Change: Class Login set DataType to password on password
Change: Class SessionInfo add field UserGrade
Change: Class FlightSearch remove field return day.

+++ src/BackOffice/Controllers/UsersController.cs
@@ -7,21 +7,24 @@ using TicketSystem.RestApiClient;
 using BackOffice.Models;
 using TicketSystem.RestApiClient.Model;
 using Microsoft.AspNetCore.Mvc.Filters;
+using Microsoft.Extensions.Configuration;
 
 namespace BackOffice.Controllers
 {
     public class UsersController : Controller
     {
+        private IConfigurationRoot config;
+        private Sessions sessions;
         private TicketApi ticketApi;
-        private ApiInformation api;
         private MessagesHandler messagesHandler;
 
         /// <summary>
         /// Default constructor, prepare api
         /// </summary>
-        public UsersController()
+        public UsersController(IConfigurationRoot newConfig, Sessions newSessions)
         {
-            api = new ApiInformation();
+            config = newConfig;
+            sessions = newSessions;
         }
 
         /// <summary>
@@ -30,16 +33,21 @@ namespace BackOffice.Controllers
         /// <param name="context">Action context</param>
         public override void OnActionExecuting(ActionExecutingContext context)
         {
-            if (TempData["Userid"] != null)
+            sessions.Intialize(context.HttpContext);
+            ViewData["Sessions"] = sessions;
+
+            if (sessions.Get("UserId") != null)
             {
-                ticketApi = new TicketApi(api.Key, api.Secret, (int)TempData["SessionId"], (string)TempData["SessionSecret"]);
+                ticketApi = new TicketApi(config["Api:Key"], config["Api:Secret"], (int)sessions.Get("SessionId"), (string)sessions.Get("SessionSecret"));
             }
             else
             {
-                ticketApi = new TicketApi(api.Key, api.Secret);
+                ticketApi = new TicketApi(config["Api:Key"], config["Api:Secret"]);
             }
 
-            messagesHandler = new MessagesHandler(TempData);
+            messagesHandler = new MessagesHandler(sessions);
+
+            ViewData["Messages"] = messagesHandler;
 
             // Execute action
             base.OnActionExecuting(context);
@@ -51,16 +59,26 @@ namespace BackOffice.Controllers
         /// <returns>View</returns>
         public IActionResult Index()
         {
-            try
+            if (sessions.Exist("UserId"))
             {
-                List<User> users = ticketApi.GetUsers();
+                List<User> users;
 
-                return View(users);
+                try
+                {
+                    users = ticketApi.GetUsers();
+                }
+                catch (Exception ex)
+                {
+                    messagesHandler.Add("danger", ex.Message);
+                    users = new List<User>();
+                }
+
+                ViewBag.Users = users;
+
+                return View();
             }
-            catch(Exception ex)
-            {
-                messagesHandler.Add("danger", ex.Message);
-            }
+
+            messagesHandler.Add("warning", "You need to be login, to see all users");
 
             return RedirectToAction("Index", "Home");
         }
@@ -70,29 +88,34 @@ namespace BackOffice.Controllers
         /// </summary>
         /// <param name="id">User index</param>
         /// <returns>View</returns>
-        [HttpGet("{id}")]
         public IActionResult Edit(string id)
         {
-            int userId;
-            if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
+            if (sessions.Exist("UserId"))
             {
-                messagesHandler.Add("danger", "No user find, to edit");
-            }
-            else
-            {
-                try
+                int userId;
+                if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
                 {
-                    User user = ticketApi.GetUser(userId);
-                    return View(user);
+                    messagesHandler.Add("danger", "No user find, to edit");
                 }
-                catch (Exception ex)
+                else
                 {
-                    messagesHandler.Add("danger", ex.Message);
+                    try
+                    {
+                        User user = ticketApi.GetUser(userId);
+                        return View(user);
+                    }
+                    catch (Exception ex)
+                    {
+                        messagesHandler.Add("danger", ex.Message);
+                    }
                 }
+
+                return RedirectToAction("Index");
             }
 
+            messagesHandler.Add("warning", "You need to be login, to edit a user");
 
-            return RedirectToAction("Index");
+            return RedirectToAction("Index", "Home");
         }
 
         /// <summary>
@@ -101,38 +124,52 @@ namespace BackOffice.Controllers
         /// <param name="id">User index</param>
         /// <param name="newUser">User object with post information</param>
         /// <returns>View or redirect</returns>
-        [HttpPost("{id}")]
+        [HttpPost]
         public IActionResult Edit(string id, User user)
         {
-            int userId;
-            if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
+            if (sessions.Exist("UserId"))
             {
-                messagesHandler.Add("danger", "No user find, to edit");
-
-                return RedirectToAction("Index");
-            }
-
-            if (user == null)
-            {
-                return View(new User());
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
+                int userId;
+                if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
                 {
-                    ticketApi.PutUser(userId, user);
-                    messagesHandler.Add("success", "User are successfully edit!");
+                    messagesHandler.Add("danger", "No user find, to edit");
 
-                    return RedirectToAction("Edit", new { id = userId });
+                    return RedirectToAction("Index");
                 }
-                catch (Exception ex)
+
+                if (user == null)
                 {
-                    messagesHandler.Add("danger", ex.Message);
+                    try
+                    {
+                        user = ticketApi.GetUser(userId);
+                        return View(user);
+                    }
+                    catch (Exception ex)
+                    {
+                        messagesHandler.Add("danger", ex.Message);
+                    }
                 }
+                else if (ModelState.IsValid)
+                {
+                    try
+                    {
+                        ticketApi.PutUser(userId, user);
+                        messagesHandler.Add("success", "User are successfully edit!");
+
+                        return RedirectToAction("Edit", new { id = userId });
+                    }
+                    catch (Exception ex)
+                    {
+                        messagesHandler.Add("danger", ex.Message);
+                    }
+                }
+
+                return View(user);
             }
 
-            return View(user);
+            messagesHandler.Add("warning", "You need to be login, to edit a user");
+
+            return RedirectToAction("Index", "Home");
         }
 
         /// <summary>
@@ -140,29 +177,35 @@ namespace BackOffice.Controllers
         /// </summary>
         /// <param name="id">User index</param>
         /// <returns>View</returns>
-        [HttpGet("{id}")]
         public IActionResult Remove(string id)
         {
-            int userId;
-            if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
+            if (sessions.Exist("UserId"))
             {
-                messagesHandler.Add("danger", "No user find, to remove");
-            }
-            else
-            {
-                try
+                int userId;
+                if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
                 {
-                    User user = ticketApi.GetUser(userId);
-
-                    return View(user);
+                    messagesHandler.Add("danger", "No user find, to remove");
                 }
-                catch (Exception ex)
+                else
                 {
-                    messagesHandler.Add("danger", ex.Message);
+                    try
+                    {
+                        User user = ticketApi.GetUser(userId);
+
+                        return View(user);
+                    }
+                    catch (Exception ex)
+                    {
+                        messagesHandler.Add("danger", ex.Message);
+                    }
                 }
+
+                return RedirectToAction("Index");
             }
 
-            return RedirectToAction("Index");
+            messagesHandler.Add("warning", "You need to be login, to remove user");
+
+            return RedirectToAction("Index", "Home");
         }
 
         /// <summary>
@@ -171,52 +214,59 @@ namespace BackOffice.Controllers
         /// <param name="id">User index</param>
         /// <param name="remove">Bool-value as string</param>
         /// <returns>Redirect or view</returns>
-        [HttpPost("{id}")]
-        public IActionResult Remove(string id, string remove = "false")
+        [HttpPost]
+        public IActionResult Remove(string id, string remove)
         {
-            int userId;
-            if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
+            if (sessions.Exist("UserId"))
             {
-                messagesHandler.Add("danger", "No user find, to remove");
-            }
-            else
-            {
-                if (remove == "true")
+                int userId;
+                if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
                 {
-                    try
-                    {
-                        ticketApi.DeleteUser(userId);
-                        messagesHandler.Add("success", "User was successfully removed!");
-
-                        return RedirectToAction("Index");
-                    }
-                    catch (Exception ex)
-                    {
-                        messagesHandler.Add("danger", ex.Message);
-                    }
+                    messagesHandler.Add("danger", "No user find, to remove");
                 }
-
-                if (remove != "false")
+                else
                 {
-                    try
+                    if (remove == "true")
                     {
-                        User user = ticketApi.GetUser(userId);
-
-                        if (remove != "true")
+                        try
                         {
-                            messagesHandler.Add("warning", "Something went wrong!");
-                        }
+                            ticketApi.DeleteUser(userId);
+                            messagesHandler.Add("success", "User was successfully removed!");
 
-                        return View(user);
+                            return RedirectToAction("Index");
+                        }
+                        catch (Exception ex)
+                        {
+                            messagesHandler.Add("danger", ex.Message);
+                        }
                     }
-                    catch(Exception ex)
+
+                    if (remove != "false")
                     {
-                        messagesHandler.Add("danger", ex.Message);
+                        try
+                        {
+                            User user = ticketApi.GetUser(userId);
+
+                            if (remove != "true")
+                            {
+                                messagesHandler.Add("warning", "Something went wrong!");
+                            }
+
+                            return View(user);
+                        }
+                        catch (Exception ex)
+                        {
+                            messagesHandler.Add("danger", ex.Message);
+                        }
                     }
                 }
+
+                return RedirectToAction("Index");
             }
 
-            return RedirectToAction("Index");
+            messagesHandler.Add("warning", "You need to be login, to remove user");
+
+            return RedirectToAction("Index", "Home");
         }
 
         /// <summary>
@@ -224,28 +274,34 @@ namespace BackOffice.Controllers
         /// </summary>
         /// <param name="id">User index</param>
         /// <returns>View</returns>
-        [HttpGet("{id}")]
         public IActionResult Tickets(string id)
         {
-            int userId;
-            if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
+            if (sessions.Exist("UserId"))
             {
-                messagesHandler.Add("danger", "No user find, to get tickets for");
-            }
-            else
-            {
-                try
+                int userId;
+                if ((String.IsNullOrEmpty(id)) || !int.TryParse(id, out userId))
                 {
-                    List<Ticket> tickets = ticketApi.GetTicketsByUser(userId);
-                    return View(tickets);
+                    messagesHandler.Add("danger", "No user find, to get tickets for");
                 }
-                catch (Exception ex)
+                else
                 {
-                    messagesHandler.Add("danger", ex.Message);
+                    try
+                    {
+                        ViewBag.Tickets = ticketApi.GetTicketsByUser(userId);
+                        return View();
+                    }
+                    catch (Exception ex)
+                    {
+                        messagesHandler.Add("danger", ex.Message);
+                    }
                 }
+
+                return RedirectToAction("Index");
             }
 
-            return RedirectToAction("Index");
+            messagesHandler.Add("warning", "You need to be login, to see user's tickets");
+
+            return RedirectToAction("Index", "Home");
         }
     }
 }
